Master source code that will compile on visual studio 2015

+++ source/mpgame/SecurityCamera.cpp
@@ -119,7 +119,7 @@ void idSecurityCamera::Spawn( void ) {
 	}
 
 	negativeSweep = ( sweepAngle < 0 ) ? true : false;
-	sweepAngle = abs( sweepAngle );
+	sweepAngle = abs( (int)sweepAngle );
 
 	scanFovCos = idMath::Cos( scanFov * idMath::PI / 360.0f );
